Add files via upload
Game of Fifteen written, mostly, in C++ with several new features. This code is optimized for code-blocks.

BEFORE YOU RUN THIS CODE:

After hitting 'run', please maximize your code block’s output window for intended
and best results. 




PROBLEM DESCRIPTION:

Game of Fifteen is a challenging game. The idea of Game of Fifteen is not new; it
has mesmerized mathematical geniuses for centuries and it will mesmerize you today.
In this puzzle, user has to arrange tiles of randomly assorted numbers in ascending
order. 

1- Game of Fifteen has variable difficulty levels (3 to 9)
2- To solve Game of Fifteen, you will have to arrange tiles, initially placed randomly, in ascending order.
3- Enter the number of tile from keyboard and that tile will move to "empty space".

4- ONE TILE AT A TIME, ADJACENT TO EMPTY SPACE, IS MOVEABLE.

5- A tile can neither jump over other tiles nor it can move diagonally to empty space. 
6- To make things juicier, I decided to limit user